+++ hrms_dashboard/static/src/js/dashboard.js
@@ -52,19 +52,49 @@ export class HrDashboard extends Component{
                 this.state.upcoming_events = res['event'];
                 this.state.announcements = res['announcement'];
             }
+            var projectTaskDetails = await this.orm.call('hr.employee', 'get_employee_project_tasks', [])
+            if (projectTaskDetails) {
+                this.state.login_employee['project_task_lines'] = projectTaskDetails;
+            }
         });
         onMounted(() => {
             this.title = 'Dashboard'
             this.render_graphs();
         });
     }
+    add_project_task() {
+                this.action.doAction({
+                    name: _t("Project Task"),
+                    type: 'ir.actions.act_window',
+                    res_model: 'project.task',
+                    view_mode: 'form',
+                    views: [[false, 'form']],
+                    target: 'new',
+                    context: {
+                        'default_user_ids': [session.uid]
+                    }
+                });
+            }
+    view_project_tasks() {
+                this.action.doAction({
+                    name: _t("My Tasks"),
+                    type: 'ir.actions.act_window',
+                    res_model: 'project.task',
+                    view_mode: 'tree,form,kanban',
+                    views: [[false, 'list'],[false, 'form'],[false, 'kanban']],
+                    domain: [['user_ids','in', session.uid]],
+                    target: 'current'
+                });
+            }
     render_graphs(){
         var self = this;
         if (this.state.login_employee){
-            self.render_department_employee();
-            self.render_leave_graph();
-            self.update_join_resign_trends();
-            self.update_monthly_attrition();
+            if (this.state.is_manager) {
+             self.render_department_employee();
+                self.render_leave_graph();
+                self.update_join_resign_trends();
+                self.update_monthly_attrition();
+            }
             self.update_leave_trend();
             self.render_employee_skill();
         }